+++ packages/vitest/src/index.ts
@@ -1,7 +1,6 @@
 import { startVitest, parseCLI } from 'vitest/node';
 import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
 import tsconfigPaths from 'vite-tsconfig-paths';
-import { cjsInterop } from 'vite-plugin-cjs-interop';
 
 type SkuContext = {
   cjsInteropDependencies: string[];
@@ -20,28 +19,24 @@ export const runVitest = async ({
   const results = parseCLI(['vitest', ...args]);
   const { cjsInteropDependencies, compilePackages } = skuContext;
 
-  const vitest = await startVitest(
+  const ctx = await startVitest(
     'test',
     results.filter,
     { config: false, ...results.options },
     {
-      plugins: [
-        vanillaExtractPlugin(),
-        tsconfigPaths(),
-        cjsInterop({
-          dependencies: [...cjsInteropDependencies],
-        }),
-      ],
+      plugins: [vanillaExtractPlugin(), tsconfigPaths()],
       test: {
         environment: 'jsdom',
         setupFiles,
       },
       ssr: {
-        noExternal: [...compilePackages],
+        noExternal: [...compilePackages, ...cjsInteropDependencies],
       },
     },
     {},
   );
 
-  await vitest.close();
+  if (!ctx.shouldKeepServer()) {
+    await ctx.exit();
+  }
 };