+++ packages/codemod/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @sku-lib/codemod
 
+## 0.3.0
+
+### Minor Changes
+
+- Add support for converting failing tests such as `test.failing` to `test.fails` ([#1401](https://github.com/seek-oss/sku/pull/1401))
+
+- Add support for detecting chained test methods such as `it.each` and `test.skip.each` ([#1400](https://github.com/seek-oss/sku/pull/1400))
+
+- Add support for converting `jest.Mock` and `jest.MockedFunction` typecasts into `vi.mocked` function calls ([#1396](https://github.com/seek-oss/sku/pull/1396))
+
+## 0.2.0
+
+### Minor Changes
+
+- Only accept paths to directories instead of any glob expression ([#1388](https://github.com/seek-oss/sku/pull/1388))
+
+  **BREAKING CHANGE**:
+  Globs are no longer accepted as input to the CLI. Only paths to directories are accepted.
+
+- Add `jest-to-vitest` codemod ([#1388](https://github.com/seek-oss/sku/pull/1388))
+
+  The `jest-to-vitest` codemod automates the migration of test files from Jest to Vitest. The following code patterns are transformed:
+  - Renaming `jest` to `vi`
+  - Importing globals such as `expect, it, describe, beforeAll`
+  - Updating `jest.requireActual` to `await vi.importActual`
+  - Wrapping functions that call `await vi.importActual` with `async`
+
+  **EXAMPLE USAGE**:
+
+  ```sh
+  pnpm dlx @sku-lib/codemod jest-to-vitest .
+  ```
+
+### Patch Changes
+
+- Correctly calculate changed files ([#1388](https://github.com/seek-oss/sku/pull/1388))
+
+- Restrict files to only those with JavaScript and TypeScript extensions ([#1388](https://github.com/seek-oss/sku/pull/1388))
+
 ## 0.1.1
 
 ### Patch Changes