+++ packages/sku/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sku
 
+## 14.12.2
+
+### Patch Changes
+
+- `vitest`: Correctly forward compile packages ([#1397](https://github.com/seek-oss/sku/pull/1397))
+
+## 14.12.1
+
+### Patch Changes
+
+- `vitest`: Correctly forward CJS interop dependencies ([#1395](https://github.com/seek-oss/sku/pull/1395))
+
 ## 14.12.0
 
 ### Minor Changes